# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Helix API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/prompts/": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "List Prompts",
        "description": "List all registered prompts.",
        "operationId": "list_prompts_api_prompts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PromptSummary"
                  },
                  "type": "array",
                  "title": "Response List Prompts Api Prompts  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "prompts"
        ],
        "summary": "Create Prompt",
        "description": "Register a new prompt.",
        "operationId": "create_prompt_api_prompts__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}/config": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get Prompt Config",
        "description": "Return prompt config with provenance info.\n\nReads PromptConfig from DB. Returns both the effective (merged) config\nper role and the raw overrides dict, so the UI can show Global/Override badges.",
        "operationId": "get_prompt_config_api_prompts__prompt_id__config_get",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "prompts"
        ],
        "summary": "Update Prompt Config",
        "description": "Update per-prompt config in DB PromptConfig table.\n\nUpserts the PromptConfig row. Stores role-prefixed fields in the extra\nJSON column. Does NOT write config.json.",
        "operationId": "update_prompt_config_api_prompts__prompt_id__config_put",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptConfigSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get Prompt",
        "description": "Get a prompt's full detail including template text from DB.",
        "operationId": "get_prompt_api_prompts__prompt_id__get",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}/template": {
      "put": {
        "tags": [
          "prompts"
        ],
        "summary": "Update Template",
        "description": "Update a prompt's template text.",
        "operationId": "update_template_api_prompts__prompt_id__template_put",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}/versions": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "List Versions",
        "description": "List all versions for a prompt, ordered by version number.",
        "operationId": "list_versions_api_prompts__prompt_id__versions_get",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PromptVersionResponse"
                  },
                  "title": "Response List Versions Api Prompts  Prompt Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}/versions/{version}": {
      "get": {
        "tags": [
          "prompts"
        ],
        "summary": "Get Version",
        "description": "Get a specific version of a prompt.",
        "operationId": "get_version_api_prompts__prompt_id__versions__version__get",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}/versions/{version}/activate": {
      "put": {
        "tags": [
          "prompts"
        ],
        "summary": "Activate Version",
        "description": "Activate a specific version for a prompt.",
        "operationId": "activate_version_api_prompts__prompt_id__versions__version__activate_put",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prompts/{prompt_id}/versions/accept": {
      "post": {
        "tags": [
          "prompts"
        ],
        "summary": "Accept Version",
        "description": "Accept an evolved template as a new version.\n\nCreates a new version from the provided template and sets it as active.\nUsed after evolution runs to accept the best evolved template.",
        "operationId": "accept_version_api_prompts__prompt_id__versions_accept_post",
        "parameters": [
          {
            "name": "prompt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prompt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptVersionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePromptRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "template": {
            "type": "string",
            "title": "Template"
          },
          "variables": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variables"
          },
          "tools": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tools"
          },
          "tool_schemas": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Schemas"
          },
          "mocks": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mocks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "purpose",
          "template"
        ],
        "title": "CreatePromptRequest",
        "description": "Request body for registering a new prompt."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PromptConfigResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/RoleConfigResponse"
          },
          "target": {
            "$ref": "#/components/schemas/RoleConfigResponse"
          },
          "judge": {
            "$ref": "#/components/schemas/RoleConfigResponse"
          },
          "tool_mocker": {
            "$ref": "#/components/schemas/ToolMockerConfigResponse"
          },
          "overrides": {
            "additionalProperties": true,
            "type": "object",
            "title": "Overrides"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "target",
          "judge",
          "tool_mocker",
          "overrides"
        ],
        "title": "PromptConfigResponse",
        "description": "Response from GET/PUT /api/prompts/{id}/config.\n\nContains effective (merged) values per role and the raw overrides dict\nso the UI can show Global/Override provenance badges."
      },
...
      "PromptDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "template_variables": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Template Variables"
          },
          "anchor_variables": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Anchor Variables"
          },
          "template": {
            "type": "string",
            "title": "Template"
          },
          "tools": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tools"
          },
          "tool_schemas": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Schemas"
          },
          "mocks": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mocks"
          },
          "variable_definitions": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variable Definitions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "purpose",
          "template_variables",
          "anchor_variables",
          "template"
        ],
        "title": "PromptDetail",
        "description": "Full prompt detail including template content."
      },
      "PromptSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "template_variables": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Template Variables"
          },
          "anchor_variables": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Anchor Variables"
          }
        },
        "type": "object",
        "required": [
          "id",
          "purpose",
          "template_variables",
          "anchor_variables"
        ],
        "title": "PromptSummary",
        "description": "Summary representation of a registered prompt."
      },
...
    }
  }
}